fix dispmanx build
For when the user is building "--with-flavors dispmanx-glesv2", fixes a bunch
of "redeclared as different kind of symbol" problems.

This patch is almost entirely from:
	https://github.com/glmark2/glmark2/commit/3b1a0f78d03ae126b78ee52a7c434809115e5993

Which was created by: Alexandros Frantzis <alexandros.frantzis@collabora.com>

See:
	https://github.com/glmark2/glmark2/issues/80

My contribution is to tweak the wscript to link all the required libraries.

Upstream-status: submitted [https://github.com/glmark2/glmark2/pull/135]
Signed-off-by: Trevor Woerner <twoerner@gmail.com>

%% original patch: 0001-fix-dispmanx-build.patch